Separate ContextCommandAutoBuilder into each context command type
