+++ src/main/kotlin/io/github/freya022/botcommands/internal/commands/application/autobuilder/ContextCommandAutoBuilder.kt
@@ -1,162 +1,33 @@
 package io.github.freya022.botcommands.internal.commands.application.autobuilder
 
 import io.github.freya022.botcommands.api.commands.CommandPath
-import io.github.freya022.botcommands.api.commands.annotations.Command
 import io.github.freya022.botcommands.api.commands.annotations.GeneratedOption
 import io.github.freya022.botcommands.api.commands.application.ApplicationCommand
-import io.github.freya022.botcommands.api.commands.application.CommandScope
-import io.github.freya022.botcommands.api.commands.application.annotations.CommandId
 import io.github.freya022.botcommands.api.commands.application.builder.ApplicationCommandBuilder
 import io.github.freya022.botcommands.api.commands.application.context.annotations.ContextOption
-import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAMessageCommand
-import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAUserCommand
 import io.github.freya022.botcommands.api.commands.application.context.builder.MessageCommandBuilder
 import io.github.freya022.botcommands.api.commands.application.context.builder.UserCommandBuilder
-import io.github.freya022.botcommands.api.commands.application.context.message.GlobalMessageEvent
-import io.github.freya022.botcommands.api.commands.application.context.user.GlobalUserEvent
-import io.github.freya022.botcommands.api.commands.application.declaration.AbstractApplicationCommandManager
-import io.github.freya022.botcommands.api.commands.application.declaration.GlobalApplicationCommandManager
-import io.github.freya022.botcommands.api.commands.application.declaration.GuildApplicationCommandManager
 import io.github.freya022.botcommands.api.core.reflect.ParameterType
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.parameters.ResolverContainer
-import io.github.freya022.botcommands.internal.commands.SkipLogger
-import io.github.freya022.botcommands.internal.commands.application.autobuilder.metadata.MessageContextFunctionMetadata
-import io.github.freya022.botcommands.internal.commands.application.autobuilder.metadata.UserContextFunctionMetadata
-import io.github.freya022.botcommands.internal.commands.autobuilder.*
+import io.github.freya022.botcommands.internal.commands.autobuilder.requireCustomOption
 import io.github.freya022.botcommands.internal.core.BContextImpl
-import io.github.freya022.botcommands.internal.core.requiredFilter
-import io.github.freya022.botcommands.internal.utils.*
 import io.github.freya022.botcommands.internal.utils.ReflectionUtils.nonInstanceParameters
-import io.github.oshai.kotlinlogging.KotlinLogging
+import io.github.freya022.botcommands.internal.utils.findDeclarationName
+import io.github.freya022.botcommands.internal.utils.findOptionName
 import net.dv8tion.jda.api.entities.Guild
 import kotlin.reflect.KFunction
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.hasAnnotation
-import net.dv8tion.jda.api.interactions.commands.Command as JDACommand
-
-private val logger = KotlinLogging.logger { }
 
 @BService
-internal class ContextCommandAutoBuilder(
+internal sealed class ContextCommandAutoBuilder(
     context: BContextImpl,
     private val resolverContainer: ResolverContainer
 ) {
-    private val forceGuildCommands = context.applicationConfig.forceGuildCommands
+    protected val forceGuildCommands = context.applicationConfig.forceGuildCommands
 
-    private val messageFunctions: List<MessageContextFunctionMetadata>
-    private val userFunctions: List<UserContextFunctionMetadata>
-
-    init {
-        messageFunctions = context.instantiableServiceAnnotationsMap
-            .getInstantiableFunctionsWithAnnotation<Command, JDAMessageCommand>()
-            .requiredFilter(FunctionFilter.nonStatic())
-            .requiredFilter(FunctionFilter.firstArg(GlobalMessageEvent::class))
-            .map {
-                val func = it.function
-                val annotation = func.findAnnotation<JDAMessageCommand>() ?: throwInternal("${annotationRef<JDAMessageCommand>()} should be present")
-                val path = CommandPath.ofName(annotation.name)
-                val commandId = func.findAnnotation<CommandId>()?.value
-
-                MessageContextFunctionMetadata(it, annotation, path, commandId)
-            }
-
-        userFunctions = context.instantiableServiceAnnotationsMap
-            .getInstantiableFunctionsWithAnnotation<Command, JDAUserCommand>()
-            .requiredFilter(FunctionFilter.nonStatic())
-            .requiredFilter(FunctionFilter.firstArg(GlobalUserEvent::class))
-            .map {
-                val func = it.function
-                val annotation = func.findAnnotation<JDAUserCommand>() ?: throwInternal("${annotationRef<JDAUserCommand>()} should be present")
-                val path = CommandPath.ofName(annotation.name)
-                val commandId = func.findAnnotation<CommandId>()?.value
-
-                UserContextFunctionMetadata(it, annotation, path, commandId)
-            }
-    }
-
-    //Separated functions so message errors don't prevent user commands from being registered
-    fun declareGlobalMessage(manager: GlobalApplicationCommandManager) = declareMessage(manager)
-
-    fun declareGlobalUser(manager: GlobalApplicationCommandManager) = declareUser(manager)
-
-    fun declareGuildMessage(manager: GuildApplicationCommandManager) = declareMessage(manager)
-
-    fun declareGuildUser(manager: GuildApplicationCommandManager) = declareUser(manager)
-
-    private fun declareMessage(manager: AbstractApplicationCommandManager) {
-        val skipLogger = SkipLogger(logger)
-        messageFunctions.forEachWithDelayedExceptions { metadata ->
-            runFiltered(
-                manager,
-                skipLogger,
-                forceGuildCommands,
-                metadata.path,
-                metadata.instance,
-                metadata.commandId,
-                metadata.func,
-                metadata.annotation.scope
-            ) { processMessageCommand(manager, metadata) }
-        }
-        skipLogger.log((manager as? GuildApplicationCommandManager)?.guild, JDACommand.Type.MESSAGE)
-    }
-
-    private fun declareUser(manager: AbstractApplicationCommandManager) {
-        val skipLogger = SkipLogger(logger)
-        userFunctions.forEachWithDelayedExceptions { metadata ->
-            runFiltered(
-                manager,
-                skipLogger,
-                forceGuildCommands,
-                metadata.path,
-                metadata.instance,
-                metadata.commandId,
-                metadata.func,
-                metadata.annotation.scope
-            ) { processUserCommand(manager, metadata) }
-        }
-        skipLogger.log((manager as? GuildApplicationCommandManager)?.guild, JDACommand.Type.USER)
-    }
-
-    private fun processMessageCommand(manager: AbstractApplicationCommandManager, metadata: MessageContextFunctionMetadata) {
-        val func = metadata.func
-        val instance = metadata.instance
-        val path = metadata.path
-        val commandId = metadata.commandId
-
-        val annotation = metadata.annotation
-        val actualScope = if (forceGuildCommands) CommandScope.GUILD else annotation.scope
-        manager.messageCommand(path.name, actualScope, func.castFunction()) {
-            fillCommandBuilder(func)
-            fillApplicationCommandBuilder(func)
-
-            isDefaultLocked = annotation.defaultLocked
-            nsfw = annotation.nsfw
-
-            processOptions((manager as? GuildApplicationCommandManager)?.guild, func, instance, commandId)
-        }
-    }
-
-    private fun processUserCommand(manager: AbstractApplicationCommandManager, metadata: UserContextFunctionMetadata) {
-        val func = metadata.func
-        val instance = metadata.instance
-        val path = metadata.path
-        val commandId = metadata.commandId
-
-        val annotation = metadata.annotation
-        val actualScope = if (forceGuildCommands) CommandScope.GUILD else annotation.scope
-        manager.userCommand(path.name, actualScope, func.castFunction()) {
-            fillCommandBuilder(func)
-            fillApplicationCommandBuilder(func)
-
-            isDefaultLocked = annotation.defaultLocked
-            nsfw = annotation.nsfw
-
-            processOptions((manager as? GuildApplicationCommandManager)?.guild, func, instance, commandId)
-        }
-    }
-
-    private fun ApplicationCommandBuilder<*>.processOptions(
+    protected fun ApplicationCommandBuilder<*>.processOptions(
         guild: Guild?,
         func: KFunction<*>,
         instance: ApplicationCommand,
